Remove minimap; add Addons hub (JourneyMap) and move BQu controls to Settings
- Remove MinimapHUD, its N keybind, showMinimap default, minimapBackground color
- Add extensible client AddonPanelRegistry + AddonsPanel opened from the party menu
- Add JourneyMapAddonPanel with a claim-overlay visibility toggle (JMapClientConfig)
- Consolidate BQu link/unlink toggle + native-manager shortcut into SettingsPanel
  (Party Info tab, shown only when BQu is present)
- Update en_us/ja_jp lang (blpc.addons.*) and blpc-overview architecture docs
