+++ src/main/java/com/github/gtexpert/blpc/integration/jmap/JMapModule.java
@@ -15,6 +15,7 @@ import org.jetbrains.annotations.NotNull;
 import com.github.gtexpert.blpc.Tags;
 import com.github.gtexpert.blpc.api.modules.TModule;
 import com.github.gtexpert.blpc.api.util.Mods;
+import com.github.gtexpert.blpc.client.gui.addons.AddonPanelRegistry;
 import com.github.gtexpert.blpc.integration.IntegrationSubmodule;
 import com.github.gtexpert.blpc.module.Modules;
 
@@ -43,6 +44,8 @@ public class JMapModule extends IntegrationSubmodule {
             syncHandler = new JMapClaimSyncHandler();
             syncHandler.register();
             MinecraftForge.EVENT_BUS.register(this);
+            AddonPanelRegistry.register(
+                    "blpc.addons.journeymap", null, () -> true, JourneyMapAddonPanel::build);
         }
     }
