+++ src/main/resources/assets/blpc/lang/ja_jp.lang
@@ -7,7 +7,6 @@ config.blpc.protection=保護
 
 # Key bindings
 key.blpc.open_map=チャンクマップを開く
-key.blpc.toggle=ミニマップ切替
 key.categories.blpc=Better Link Party Claim
 
 # Commands
@@ -105,12 +104,12 @@ blpc.party.trust_level.owner=Owner
 # 保護設定
 blpc.party.settings_protection=-- 保護設定 --
 blpc.party.settings_allies=-- 同盟 --
-blpc.party.settings_enemies=-- 敵対 --
+blpc.party.settings_enemies=-- 敌対 --
 blpc.party.settings_party=-- パーティ情報 --
 blpc.party.settings_tab_party=パーティ情報
 blpc.party.settings_tab_protection=保護設定
 blpc.party.settings_tab_allies=同盟
-blpc.party.settings_tab_enemies=敵対
+blpc.party.settings_tab_enemies=敌対
 blpc.party.tab.parties=パーティー
 blpc.party.tab.players=プレイヤー
 blpc.party.tab.no_party=パーティー無し
@@ -120,9 +119,9 @@ blpc.party.no_other_parties=他のパーティーが見つかりません
 blpc.party.no_players_online=オンラインプレイヤーがいません
 blpc.party.search=検索...
 blpc.party.tooltip.toggle_ally=クリックで同盟追加/解除
-blpc.party.tooltip.toggle_enemy=クリックで敵対追加/解除
+blpc.party.tooltip.toggle_enemy=クリックで敌対追加/解除
 blpc.party.add_ally=+ 同盟パーティー追加
-blpc.party.add_enemy=+ 敵対パーティー追加
+blpc.party.add_enemy=+ 敌対パーティー追加
 blpc.party.fakeplayer_trust=Fake Playerの権限
 blpc.party.free_to_join_on=自由参加: ON
 blpc.party.free_to_join_off=自由参加: OFF
@@ -139,7 +138,7 @@ blpc.party.invited_label=招待中
 blpc.party.tooltip.accept_invite=クリックで招待を承認
 blpc.party.tooltip.join_free=クリックでこのパーティに参加
 
-# 味方 / 敵対
+# 味方 / 敌対
 blpc.party.add=追加
 blpc.party.remove=削除
 blpc.party.promote=昇格
@@ -156,7 +155,7 @@ blpc.party.transfer_confirm=移譲
 blpc.party.tooltip.free_to_join=招待なしで誰でも参加できます
 blpc.party.tooltip.max_members=パーティの最大メンバー数 (0 = 無制限)
 blpc.party.tooltip.color=チャンクマップ上のチーム表示色
-blpc.party.tooltip.description=パーティ名の下に表示されます
+blpc.party.tooltip.description=パーティー名の下に表示されます
 blpc.party.tooltip.explosion=保護チャンク内の爆発を許可または禁止します
 blpc.party.tooltip.fakeplayer=自動化MOD(BuildCraft, EnderIO等)の権限レベル
 blpc.party.tooltip.trust_level=この操作に必要な最低権限レベル
@@ -165,9 +164,9 @@ blpc.party.tooltip.kick=クリックでキック
 blpc.party.tooltip.invite=クリックで招待
 blpc.party.tooltip.member_self=クリックでパーティを脱退
 blpc.party.tooltip.ally=クリックで味方ステータスを切替
-blpc.party.tooltip.enemy=クリックで敵対ステータスを切替
+blpc.party.tooltip.enemy=クリックで敌対ステータスを切替
 blpc.party.tooltip.remove_ally=クリックで同盟解除
-blpc.party.tooltip.remove_enemy=クリックで敵対解除
+blpc.party.tooltip.remove_enemy=クリックで敌対解除
 blpc.party.tooltip.moderator=クリックで昇格 / 降格
 blpc.party.tooltip.default=デフォルト値: %s
 blpc.party.tooltip.options=選択肢:
@@ -175,14 +174,26 @@ blpc.party.tooltip.settings=保護とパーティー設定
 blpc.party.tooltip.members=パーティーメンバーを表示
 blpc.party.tooltip.allies=同盟パーティーの管理
 blpc.party.tooltip.moderators=モデレーターを表示
-blpc.party.tooltip.enemies=敵対パーティーの管理
+blpc.party.tooltip.enemies=敌対パーティーの管理
 blpc.party.tooltip.transfer=他のメンバーにオーナー権を譲渡
 blpc.party.dialog.add_ally=同盟パーティーを選択
-blpc.party.dialog.add_enemy=敵対パーティーを選択
+blpc.party.dialog.add_enemy=敌対パーティーを選択
 blpc.party.tooltip.open_native=BetterQuestingのパーティー管理を開く
 blpc.party.tooltip.link_bqu=BQuパーティ連携を切替\nBQuパーティが必要です
 blpc.party.tooltip.bqu_no_party=BQuパーティが見つかりません
 blpc.party.tooltip.bqu_party_info=BQuパーティ
+
+# アドオン
+blpc.addons=アドオン
+blpc.addons.title=アドオン
+blpc.addons.tooltip=他Modとの連携設定
+blpc.addons.empty=利用可能なアドオンはありません
+blpc.addons.journeymap=JourneyMap
+blpc.addons.journeymap.title=JourneyMap
+blpc.addons.journeymap.overlays_on=クレーム表示: ON
+blpc.addons.journeymap.overlays_off=クレーム表示: OFF
+blpc.addons.journeymap.overlays_tooltip=JourneyMap上にBLPCのクレーム範囲を表示
+blpc.addons.journeymap.waypoints_soon=チームウェイポイント共有（準備中）
 
 # 領域出入り通知
 blpc.transit.member.enter=%sが帰宅した
